+++ libraries/Bluefruit52Lib/src/BLEGatt.h
@@ -49,7 +49,7 @@
 #define BLE_GATT_MAX_SERVER_CHARS        40
 
 #define BLE_GATT_MAX_CLIENT_SERVICE      20
-#define BLE_GATT_MAX_CLIENT_CHARS        30
+#define BLE_GATT_MAX_CLIENT_CHARS        40
 
 
 class BLEGatt